Adding a dateObject to the ride interface
Added a dateObject that is the ISO

(2019-03-27T07:12:00.000Z)

of the actual input for the date and time
given when adding the ride.

Next: convert the ISO into a date with new Date(iso);
and then add the ability front and backend to sort by
the soonest date, which is easy with built in methods
in javascript.