Creating filter box
Creating the filter box as a placeholder for the complete
implementation of filtering the date and time of the ride.

To move on, we need to figure out how to input a ISO time
from the client and correctly send that in to the filtering
methods that exist.

Addresses issue #5