Filtering Complete
Added a button on the filter box that calls
refreshRides() as well as changing the way the
rides are called in the html.

This allows us to actually filter rides by their
departure date.

Next we need to continue to display the rest of the
rides, but in order of closest departure.

Addresses issue #5 #6