Implement a server-side token generation scheme
The provided method for creating a user token created a read-access token only. I want to be able to make all GetStream API calls client-side, so having write/modify-access is needed. Using GetStream's docs and the auth0.jwt library, I created the method in ChatRequestHandler 'authenticateDevUser' that generates an all-access JWT that GetStream will recognize.

We will eventually want to grant tokens that only give users access to certain feeds, but this will do for now!