Beginning Ordered Departure
Added functionality that sorts rides that have been added in
the client to sort themselves by their respective dates and
times of departure. This works for all rides added in client,
but does not work for rides that were seeded into the database.
I'm not sure why this is.

Along with that, filtering by a date does not show rides that
were added on that date before 5 AM. I believe this is because
the Date object automatically inserts the local timezone into the
filtered date, but not necessarily the date created when the
call is made to add a new ride.

Addresses issues #5 #6