Modify the '_id' field in ride.ts to explicitly include '$oid' + update tests & code to reflect this change
Trying to access '$oid' without having this explicitly defined was causing the build to fail. But simply removing references to $oid breaks the code. So I changed the '_id' field in ride.ts from '_id: string' to '_id: { $oid: string }' and updated references in the code to reflect this change.