Restructure to update filteredRides within innermost asynchronous call
This is a complicated commit. We realized that there are several layers of asynchronous behavior, and that any updating of the final result (filteredRides) has to be done inside the innermost async call.

We also cleaned up the code, removing commented out code and unneeded variables that had been carried over from when we had this logic inside RideListComponent.

The filter is now giving an end result that seems to be correct, but the RideListComponent that we add as a listener is not seeing it. This is our next todo.

With @koran023